Fixed UI for universal Renderer
