+++ com.unity.render-pipelines.universal/Runtime/ScriptableRendererFeature.cs
@@ -1,5 +1,5 @@
 using System;
-using UnityEngine.Scripting.APIUpdating;
+using UnityEngine;
 
 namespace UnityEngine.Rendering.Universal
 {
